implemented do_bootafter()
Field 'boot-after' in fruit.json contains a unix timestamp
(in miliseconds) that if the OS was booted before it, then
the system will be rebooted, otherwise nothing will be done.

Note that the value of 'boot-after' should not be too far
from current unix timestamp in order to avoid the system
rebooted multiple times.

Suggested value: use current unix timestamp.