fixed bug when 'apk update' exits with error
Some APK mirrors use HTTPS, but their certificates are invalid.
This makes 'apk update' to exit with error, which fails the entire
updating process and leaves the non-active root partition to be in
state dirty. There is no escape on this 'devil loop' since
fruit-update will automatically restore the dirty partition by
cloning the clean on the dirty one. However, when fruit-update
retries to update the recovered root partition, it will fail and
mark it as dirty again.

To solve this, fruit-update will ignore any error of 'apk update'
command since it only downloads and updates the local package-list.